backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gitpython==3.1.40
sqlalchemy==2.0.23
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
pandas==2.1.4
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1
